Avoid race on pidfile creation
Avoid a race on pidfile creation by creating the pidfile before the JSON is processed.
Prior to this patch, the pidfile was created only after the JSON had been processed.
This creates a race if the initiating process assumes the daemon is running if it
the call to start the daemon has returned. This corrects the problem by not consuming
the JSON (and thus preventing the parent from quitting) until the lock file has been
created.

Signed-off-by: Alex Bligh <alex@alex.org.uk>